# lds-openapi.yml
openapi: 3.0.1
info:
  title: LDS - Location Data Supplier
  version: v1.0
  description: |
    LDS - API Documentation.

    ## Changelog
    Date       | Version | Changes
    -----------|---------|----------------------------------------------------------------------------------------------------------
    xx.xx.2020 |  v1.0   | First api version.
servers:
  - url: http://localhost:40200
    description: "Generated server url"
paths:
...
  /lds/api/v1.0/locations/{id}:
    get:
      tags:
        - Locations
      operationId: GetLocationById
      summary: Get Location by id
      description: Get information about location.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
            format: int64
            description: Location identitfication number.
            example: 123
      responses:
        200:
          description: Location with requested id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiLocation'
        404:
          description: Loaction for provided id does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                errorCode: 'LoactionNotFound'
                message: 'Location (id=1) not found.'
    delete:
      tags:
        - Locations
      operationId: DeleteLocation
      summary: Deletes Location by id.
      description: Deletes location by provided id and returns status of operation.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
            format: int64
            description: Location identitfication number.
            example: 123
      responses:
        204:
          description: Successfuly deleted Location. No content.
        404:
          description: Loaction for provided id does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                errorCode: 'LoactionNotFound'
                message: 'Location (id=1) not found.'
...
components:
  schemas:
    ApiErrorResponse:
      type: object
      description: Description of error which occurred while processing api call.
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          description: Name of api error.
          example: LoactionNotFound
        errorMessage:
          type: string
          description: Detailed message about api error.
    ApiLocation:
      type: object
      description: Represents location.
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: string
          description: Locations latitude.
          example: 10.123123
        longitude:
          type: string
          description: Locations longitude.
          example: 11.123123
        altitude:
          type: string
          description: Locations altitude.
          example: 12.123123
...